+++ src/common/collections/contents/server/publish.js
@@ -8,8 +8,8 @@ const assert = require( 'assert' ).strict; // up to nodejs v16.x
 // publishes the list of known contents
 //  have the full content for each document + add a length data
 
-Meteor.publish( 'teContent.listAll', function( collection_name ){
-    assert( collection_name && _.isString( collection_name ) && collection_name.length, 'teContent.listAll expects collection_name a non-empty string, found', collection_name );
+Meteor.publish( 'pwix.Editor.p.listAll', function( collection_name ){
+    assert( collection_name && _.isString( collection_name ) && collection_name.length, 'pwix.Editor.p.listAll expects collection_name a non-empty string, found', collection_name );
     const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
     const self = this;
 
@@ -38,9 +38,9 @@ Meteor.publish( 'teContent.listAll', function( collection_name ){
 });
 
 // publishes a document by its name
-Meteor.publish( 'teContent.byName', function( collection_name, document_name ){
-    assert( collection_name && _.isString( collection_name ) && collection_name.length, 'teContent.byName expects collection_name be a non-empty string, found', collection_name );
-    assert( document_name && _.isString( document_name ) && document_name.length, 'teContent.byName expects document_name be a non-empty string, found', document_name );
+Meteor.publish( 'pwix.Editor.p.byName', function( collection_name, document_name ){
+    assert( collection_name && _.isString( collection_name ) && collection_name.length, 'pwix.Editor.p.byName expects collection_name be a non-empty string, found', collection_name );
+    assert( document_name && _.isString( document_name ) && document_name.length, 'pwix.Editor.p.byName expects document_name be a non-empty string, found', document_name );
     const collection = Editor.collections.get( collection_name, Editor.collections.Contents.schema );
     return collection.find({ name: document_name });
 });