+++ src/common/collections/contents/schema.js
@@ -8,7 +8,8 @@
  * See Contents.methods and common/init/collections.js
  */
 
-import SimpleSchema from 'meteor/aldeed:simple-schema';
+import SimpleSchema from 'simpl-schema';
+
 import { Tracker } from 'meteor/tracker';
 
 Tracker.autorun(() => {